Change endpoint for sync data with AI database when sign up (#16)
This pull request includes a change to the `AccessController` class in
the `user-service/src/modules/access/access.controller.js` file. The
change modifies the endpoint used in the `syncData` function call to
ensure that user data is sent to a more secure, private URL.

*
[`user-service/src/modules/access/access.controller.js`](diffhunk://#diff-2bd403f221ea45df31aaf5a24bb5ac1522908ba80d7dac85486176f974b4c7beL41-R41):
Changed the endpoint in the `syncData` function call from `/user/create`
to `/private/user/create` to enhance security.